products crud

+++ src/utils/file-finctions.js
@@ -3,5 +3,8 @@ import path from 'path'
 
 export const deleteFile = (filePath) => {
     let fullPath = path.resolve(filePath)
-    fs.unlinkSync(fullPath)
+    // Check if the file exists before trying to delete it
+    if (fs.existsSync(fullPath)) {
+        fs.unlinkSync(fullPath);
+    }
 }
